+++ packages/muster/src/nodes/remote/utils/do-http-request.ts
@@ -1,6 +1,5 @@
 import { Observable, ObservableLike, Observer } from '@dws/muster-observable';
-const xhr = require('xhr');
-import { Emitter } from '../../../utils/emitter';
+import axios, { AxiosPromise, AxiosResponse, CancelTokenSource } from 'axios';
 import { error, ErrorNodeDefinition } from '../../graph/error';
 
 export interface HttpRequestConfiguration {
@@ -13,30 +12,20 @@ export interface HttpRequestConfiguration {
   withCredentials: boolean;
 }
 
-interface XhrResponse {
-  body: any;
-  statusCode: number;
-  method: String;
-  headers: any;
-  url: string;
-}
-
-type XhrResult = { hasTimedOut: true } | { hasTimedOut: false; body: string | ErrorNodeDefinition };
-
-export type Callback<T> = (err: any, response: T, body: any) => void;
+const CancelToken = axios.CancelToken;
 
 export function doHttpRequest(
   config: HttpRequestConfiguration,
 ): ObservableLike<string | ErrorNodeDefinition> {
   return new Observable((observer: Observer<string | ErrorNodeDefinition>) => {
+    const cancellationToken = CancelToken.source();
     let cancelled = false;
-    const abortEmitter = new Emitter<boolean>();
     // This returns promise but we ignore it...
-    startRequestRetryLoop(config, observer, abortEmitter);
+    startRequestRetryLoop(config, observer, cancellationToken);
     return () => {
       if (cancelled) return;
       cancelled = true;
-      abortEmitter.emit(true);
+      cancellationToken.cancel();
     };
   });
 }
@@ -44,11 +33,11 @@ export function doHttpRequest(
 async function startRequestRetryLoop(
   config: HttpRequestConfiguration,
   observer: Observer<string | ErrorNodeDefinition>,
-  abortEmitter: Emitter<boolean>,
+  cancellationToken: CancelTokenSource,
 ) {
   let remainingRetries = config.numberOfRetries || 0;
   do {
-    const response = await doSingleHttpRequest(config, abortEmitter);
+    const response = await doSingleHttpRequest(config, cancellationToken);
     if (!response.hasTimedOut) {
       observer.next(response.body);
       return;
@@ -63,73 +52,54 @@ async function startRequestRetryLoop(
 
 function doSingleHttpRequest(
   config: HttpRequestConfiguration,
-  abortEmitter: Emitter<boolean>,
-): Promise<XhrResult> {
-  let resolve: (res: XhrResult) => void;
-  const resultPromise = new Promise<XhrResult>((res) => (resolve = res));
-  let disposeAbortListener: (() => void) | undefined;
-  let isAborted = false;
-  let isCompleted = false;
-  const request = startRequest(config, (err: any, res: XhrResponse, body: any) => {
-    if (isAborted) return;
-    isCompleted = true;
-    disposeAbortListener && disposeAbortListener();
-    if (res.statusCode === 0) {
-      resolve({ hasTimedOut: true });
-      return;
-    }
-    resolve({ hasTimedOut: false, body: processResponse(config, err, res, body) });
-  });
-  if (!isCompleted) {
-    disposeAbortListener = abortEmitter.listen(() => {
-      if (isCompleted) return;
-      isAborted = true;
-      disposeAbortListener && disposeAbortListener();
-      request.abort();
-      resolve({ hasTimedOut: false, body: error('The request was aborted.') });
+  cancellationToken: CancelTokenSource,
+) {
+  return startRequest(config, cancellationToken)
+    .then((response) => {
+      if (response.status === 0) {
+        return { hasTimedOut: true, body: error('A request has timed out') };
+      }
+      return { hasTimedOut: false, body: processResponse(config, response) };
+    })
+    .catch((ex) => {
+      if (axios.isCancel(ex)) {
+        return { hasTimedOut: false, body: error('The request was aborted.') };
+      }
+      return { hasTimedOut: false, body: remoteError('Network error', ex) };
     });
-    if (isAborted) {
-      disposeAbortListener();
-    }
-  }
-  return resultPromise;
 }
 
 function startRequest(
   config: HttpRequestConfiguration,
-  callback: Callback<XhrResponse>,
-): XMLHttpRequest {
-  return xhr(
-    {
-      body: config.body,
-      headers: config.headers,
-      method: 'POST',
-      timeout: config.requestTimeout,
-      url: config.url,
-      useXDR: false,
-      withCredentials: config.withCredentials,
+  cancellationToken: CancelTokenSource,
+): AxiosPromise {
+  return axios({
+    cancelToken: cancellationToken.token,
+    method: 'post',
+    url: config.url,
+    data: config.body,
+    responseType: 'text',
+    headers: {
+      'Content-Type': 'application/json',
     },
-    callback,
-  );
+  });
 }
 
 function processResponse(
   config: HttpRequestConfiguration,
-  err: any,
-  res: XhrResponse,
-  body: string,
+  res: AxiosResponse<any>,
 ): string | ErrorNodeDefinition {
-  if (err) return error('Network error');
-  if (res.statusCode < 200 || res.statusCode > 299) return remoteError(body, res);
-  if (!body) return remoteError('Invalid remote server response', res);
-  return body;
+  const data = typeof res.data === 'object' ? JSON.stringify(res.data) : res.data;
+  if (res.status < 200 || res.status > 299) return remoteError(data, res);
+  if (!data) return remoteError('Invalid remote server response', res);
+  return data;
 }
 
-function remoteError(message: string, response: XhrResponse): ErrorNodeDefinition {
+function remoteError(message: string, response: AxiosResponse<any>): ErrorNodeDefinition {
   return error(message, {
     data: {
-      url: response.url,
-      statusCode: response.statusCode,
+      url: response.config.url,
+      statusCode: response.status,
     },
   });
 }