+++ CHANGELOG.md
@@ -7,6 +7,10 @@ title: Latest Changes
 
 ## Unreleased
 
+### 🚀 New Features
+
+- Added the ability to use the Muster `remote()` node on the server side
+
 ### 🐛 Bug Fixes
 
 - Fixed `startsWith()` example ordering of pattern and subject